beeper, novy klasifikator telefonov

+++ app/src/server/server.py
@@ -1,30 +1,33 @@
 import flask
 import os
+import os.path
 import io
 from flask import jsonify
-from flask import request
+from flask import request, Response
 from flask import send_file
 import matplotlib.pyplot as plt
 from matplotlib.figure import Figure
 from pylab import *
+import cv2
+import matplotlib.dates as md
+import datetime as dt
+
 
 app = flask.Flask(__name__)
 matplotlib.rc('axes',edgecolor='white')
 
-xAxis = []
 fig = plt.figure(figsize=(8,4.5))
 
 @app.route('/upload_data', methods=['POST'])
 def upload():
     try:        
-        timestamp = request.json['timestamp']
         globalD = request.json['globalDistraction']
         phoneD = request.json['phoneDistraction']
         coffeeD = request.json['coffeeDistraction']
         drowsiness = request.json['drowsinessLevel']
 
         f = open('statistics.txt', 'a')
-        f.write(str(timestamp) + " " + str(globalD) + " " + str(phoneD) + " " + str(coffeeD) + " " + str(drowsiness) + "\n")
+        f.write(str(md.date2num(dt.datetime.now())) + " " + str(globalD) + " " + str(phoneD) + " " + str(coffeeD) + " " + str(drowsiness) + "\n")
         f.close() 
 
     except Exception as e:
@@ -36,11 +39,16 @@ def upload():
 @app.route('/clear_stats', methods=['GET'])
 def clear_stats():
     try:
-        os.remove('overall.png')
-        os.remove('phone.png')
-        os.remove('coffee.png')
-        os.remove('drowsiness.png')
-        open('statistics.txt', 'w').close()
+        if(os.path.isfile('overall.png')):
+            os.remove('overall.png')
+        if(os.path.isfile('phone.png')):
+            os.remove('phone.png')
+        if(os.path.isfile('coffee.png')):
+            os.remove('coffee.png')
+        if(os.path.isfile('drowsiness.png')):    
+            os.remove('drowsiness.png')
+        if(os.path.isfile('statistics.txt')):
+            open('statistics.txt', 'w').close()
     except Exception as e:
         return jsonify({'result': "ERROR"}), 400
 
@@ -50,13 +58,16 @@ def clear_stats():
 @app.route('/load_graph_data', methods=['GET'])
 def load_graph_data():
 
-    global xAxis
+    avgGlobal = 0.0
+    avgPhone = 0.0
+    avgCoffee = 0.0
+    avgDrowsiness = 0.0
 
-    timestamps = []
-    globalDs = []
-    phoneDs = []
-    coffeeDs = []
-    drowsinesss = []
+    timestampArray = []
+    globalDistractionArray = []
+    phoneDistractionArray = []
+    coffeeDistractionArray = []
+    drowsinessLevelArray = []
     xAxis = []
 
     with open('statistics.txt') as f:
@@ -66,11 +77,11 @@ def load_graph_data():
         while line:
             values = line.split(' ')
             
-            timestamps.append(values[0])
-            globalDs.append(float(values[1]))
-            phoneDs.append(float(values[2])*2)
-            coffeeDs.append(float(values[3])*2)
-            drowsinesss.append(float(values[4])/2)
+            timestampArray.append(float(values[0]))
+            globalDistractionArray.append(float(values[1]))
+            phoneDistractionArray.append(float(values[2])*2)
+            coffeeDistractionArray.append(float(values[3])*2)
+            drowsinessLevelArray.append(float(values[4])/2)
             xAxis.append(cnt)
 
             line = f.readline()
@@ -78,20 +89,119 @@ def load_graph_data():
 
     f.close()
 
-    draw_graph(globalDs, "DISTRACTION", "Overall distraction graph", "overall.png", round(Average(globalDs), 2))
-    draw_graph(phoneDs, "PHONE USAGE", "Phone usage graph", "phone.png", round(Average(phoneDs), 2))
-    draw_graph(coffeeDs, "COFFEE HOLDING", "Coffee holding graph", "coffee.png", round(Average(coffeeDs), 2))
-    draw_graph(drowsinesss, "DROWSINESS LEVEL", "Drowsiness level graph", "drowsiness.png", round(Average(drowsinesss), 2))
+    dates=[dt.datetime.fromtimestamp(ts) for ts in timestampArray]
 
-    return jsonify({'avg_overall': str(round(Average(globalDs), 2)),
-                    'avg_phone': str(round(Average(phoneDs), 2)),
-                    'avg_coffee': str(round(Average(coffeeDs), 2)),
-                    'avg_drowsiness': str(round(Average(drowsinesss), 2))}), 200
+    if(cnt > 500):
+        valueCount = (cnt / 100) + 1
+
+        globalDistractionArrayScaled = []
+        phoneDistractionArrayScaled = []
+        coffeeDistractionArrayScaled = []
+        drowsinessLevelArrayScaled = []
+        timestampArrayScaled = []
+        xAxisScaled = []
+        summ = 0
+        counter = 0
+        newSumm = 0
+        
+        for e in globalDistractionArray:
+            summ += e
+            counter += 1
+            if counter == valueCount:
+                globalDistractionArrayScaled.append(summ / valueCount)
+                newSumm += 1
+                xAxisScaled.append(newSumm)
+                counter = 0
+                summ = 0
+
+        if counter > 0:
+            globalDistractionArrayScaled.append(summ / valueCount)
+            newSumm += 1
+            xAxisScaled.append(newSumm)
+            counter = 0
+            summ = 0
+
+        for e in phoneDistractionArray:
+            summ += e
+            counter += 1
+            if counter == valueCount:
+                phoneDistractionArrayScaled.append(summ / valueCount)
+                counter = 0
+                summ = 0
+
+        if counter > 0:
+            phoneDistractionArrayScaled.append(summ / valueCount)
+            counter = 0
+            summ = 0
+
+        for e in coffeeDistractionArray:
+            summ += e
+            counter += 1
+            if counter == valueCount:
+                coffeeDistractionArrayScaled.append(summ / valueCount)
+                counter = 0
+                summ = 0
+
+        if counter > 0:
+            coffeeDistractionArrayScaled.append(summ / valueCount)
+            counter = 0
+            summ = 0
+
+        for e in drowsinessLevelArray:
+            summ += e
+            counter += 1
+            if counter == valueCount:
+                drowsinessLevelArrayScaled.append(summ / valueCount)
+                counter = 0
+                summ = 0
+
+        if counter > 0:
+            drowsinessLevelArrayScaled.append(summ / valueCount)
+            counter = 0
+            summ = 0
+
+        for e in timestampArray:
+            summ += e
+            counter += 1
+            if counter == valueCount:
+                timestampArrayScaled.append(summ / valueCount)
+                counter = 0
+                summ = 0
+
+        if counter > 0:
+            timestampArrayScaled.append(summ / valueCount)
+            counter = 0
+            summ = 0
+
+        avgGlobal = round(Average(globalDistractionArrayScaled), 2)
+        avgPhone = round(Average(phoneDistractionArrayScaled), 2)
+        avgCoffee = round(Average(coffeeDistractionArrayScaled), 2)
+        avgDrowsiness = round(Average(drowsinessLevelArrayScaled), 2)
+
+        draw_graph(globalDistractionArrayScaled, timestampArrayScaled, "DISTRACTION", "Overall distraction graph", "overall.png", avgGlobal)
+        draw_graph(phoneDistractionArrayScaled, timestampArrayScaled, "PHONE USAGE", "Phone usage graph", "phone.png", avgPhone)
+        draw_graph(coffeeDistractionArrayScaled, timestampArrayScaled, "COFFEE HOLDING", "Coffee holding graph", "coffee.png", avgCoffee)
+        draw_graph(drowsinessLevelArrayScaled, timestampArrayScaled, "DROWSINESS LEVEL", "Drowsiness level graph", "drowsiness.png", avgDrowsiness)
+    else:
+        avgGlobal = round(Average(globalDistractionArray), 2)
+        avgPhone = round(Average(phoneDistractionArray), 2)
+        avgCoffee = round(Average(coffeeDistractionArray), 2)
+        avgDrowsiness = round(Average(drowsinessLevelArray), 2)   
+
+        draw_graph(globalDistractionArray, timestampArray, "DISTRACTION", "Overall distraction graph", "overall.png", avgGlobal)
+        draw_graph(phoneDistractionArray, timestampArray, "PHONE USAGE", "Phone usage graph", "phone.png", avgPhone)
+        draw_graph(coffeeDistractionArray, timestampArray, "COFFEE HOLDING", "Coffee holding graph", "coffee.png", avgCoffee)
+        draw_graph(drowsinessLevelArray, timestampArray, "DROWSINESS LEVEL", "Drowsiness level graph", "drowsiness.png", avgDrowsiness)
+
+    return jsonify({'avg_overall': str(avgGlobal),
+                    'avg_phone': str(avgPhone),
+                    'avg_coffee': str(avgCoffee),
+                    'avg_drowsiness': str(avgDrowsiness)}), 200
 
 
-def draw_graph(yAxis, ylabel, title, filename, averageRate):
+def draw_graph(yAxis, xAxis, ylabel, title, filename, averageRate):
 
-    global timestamps, globalDs, phoneDs, coffeeDs, drowsinesss, xAxis, fig
+    global fig
 
     if(averageRate >= 2):
         if(averageRate < 5):
@@ -103,18 +213,22 @@ def draw_graph(yAxis, ylabel, title, filename, averageRate):
     ax = fig.add_subplot(111, facecolor='#0a0a0a')
     ax.tick_params(axis='x', colors='white')
     ax.tick_params(axis='y', colors='white')
+
+    xfmt = md.DateFormatter('%H:%M:%S')
+    plt.xticks( rotation=0 )
+    ax.xaxis.set_major_formatter(xfmt)
+
     plt.axhline(y=2, color='green', linewidth=0.5, linestyle='--')
     plt.axhline(y=5, color='yellow', linewidth=0.5, linestyle='--')
-    plt.xlabel('TIME (s)', color='white')
+    plt.xlabel('TIME', color='white')
     plt.ylim(0, 11)
     
     plt.plot(xAxis[0:len(xAxis) -1 ], yAxis[0:len(yAxis) - 1], linewidth=2, color=color)
     plt.ylabel(ylabel, color='white')
-    plt.title(title, color='white')
+    #plt.title(title, color='white')
     plt.savefig(filename, facecolor=fig.get_facecolor(), edgecolor='none')
 
     plt.clf()
-    print("ok 4")
 
 
 def Average(lst): 
@@ -123,7 +237,6 @@ def Average(lst):
 
 @app.route('/get_graph', methods=['GET'])
 def get_graph():
-
     graph_type = request.args.get('type')
     return send_file(graph_type + '.png', mimetype='image/png')     
 
@@ -131,6 +244,30 @@ def get_graph():
 @app.route('/check_availability', methods=['GET'])
 def check_availability():
     return flask.jsonify({'response': 'OK'}), 200  
+
+
+
+def generate():
+    while True:
+        video = cv2.VideoCapture('test_velky_0.mp4')
+        while True:
+            ret, frame = video.read()
+            time.sleep(0.2)
+            if frame is not None:
+                (flag, encodedImage) = cv2.imencode(".jpg", frame)
+
+                if flag:
+                    yield b'--frame\r\n' 
+                    yield b'Content-Type: image/jpeg\r\n'
+                    yield b'Content-Length: ' + str(len(encodedImage)) + b'\r\n\r\n'
+                    yield encodedImage.tobytes() + b'\r\n'
+            else: 
+                break
+
+
+@app.route("/video_feed", methods=['GET'])
+def video_feed():
+   return Response(generate(), mimetype = "multipart/x-mixed-replace; boundary=--frame")
 
 
 if __name__ == '__main__':
